Create hidingElementThatRunsAfterPageIsLoad
In this particular issue, there was Elements that were being load after the document and the window was fully load. The idea was to hide on of these new Elements. It is why I add a MutationObserver on the div that I knew was about to get new HTML on it.